+++ runtime/runtime.c
@@ -1,17 +1,27 @@
 #include <stdint.h>
 #include <stdio.h>
 
-typedef struct ByteSlice
+typedef struct Slice
 {
-    char *ptr;
+    void *ptr;
     size_t len;
-} ByteSlice;
+} Slice;
+
+size_t Len(Slice slice)
+{
+    return slice.len;
+}
 
 /// @brief Prints a slice to STDOUT.
 /// @param slice
-void Print(ByteSlice slice)
+void Print(Slice slice)
 {
     fwrite(slice.ptr, 1, slice.len, stdout);
+}
+
+void PrintInt(size_t value)
+{
+    printf("%zu", value);
 }
 
 /// @brief The application entry point.