+++ src/typechecker/types.rs
@@ -66,8 +66,16 @@ impl Slice {
 /// A type expression.
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]
 pub enum Type {
+    I8,
+    I16,
     I32,
+    I64,
+    Int,
     U8,
+    U16,
+    U32,
+    U64,
+    Uint,
     Ptr(Ptr),
     Slice(Slice),
 }
@@ -76,8 +84,16 @@ impl Type {
     /// Returns the visualized name of type.
     pub fn name(&self) -> String {
         match self {
+            Type::I8 => "i8".to_string(),
+            Type::I16 => "i16".to_string(),
             Type::I32 => "i32".to_string(),
+            Type::I64 => "i64".to_string(),
+            Type::Int => "int".to_string(),
             Type::U8 => "u8".to_string(),
+            Type::U16 => "u16".to_string(),
+            Type::U32 => "u32".to_string(),
+            Type::U64 => "u64".to_string(),
+            Type::Uint => "uint".to_string(),
             Type::Ptr(ptr) => ptr.name(),
             Type::Slice(slice) => slice.name(),
         }
@@ -94,8 +110,16 @@ impl Type {
     /// Returns the size of the type in bytes.
     pub fn size(&self, ptr_size: usize) -> usize {
         match self {
+            Type::I8 => 1,
+            Type::I16 => 2,
             Type::I32 => 4,
+            Type::I64 => 8,
+            Type::Int => ptr_size,
             Type::U8 => 1,
+            Type::U16 => 2,
+            Type::U32 => 4,
+            Type::U64 => 8,
+            Type::Uint => ptr_size,
             Type::Ptr(_) => ptr_size,
             Type::Slice(_) => ptr_size * 2,
         }
@@ -107,8 +131,16 @@ impl Type {
     pub fn check(module: &mut Scope, ty: &ast::Type) -> Result<Self, Error> {
         match ty {
             ast::Type::Named(name) => match name.value.as_str() {
+                "i8" => Ok(Type::I8),
+                "i16" => Ok(Type::I16),
                 "i32" => Ok(Type::I32),
+                "i64" => Ok(Type::I64),
+                "int" => Ok(Type::Int),
                 "u8" => Ok(Type::U8),
+                "u16" => Ok(Type::U16),
+                "u32" => Ok(Type::U32),
+                "u64" => Ok(Type::U64),
+                "uint" => Ok(Type::Uint),
                 _ => return Err(Error::UnknownNamedType(name.clone())),
             },
             ast::Type::Pointer(ptr) => {